[#12]validation 적용
