+++ src/main/resources/application.yaml
@@ -19,6 +19,6 @@ mybatis:
 
 logging:
   level:
-    org.springframework: info
+    org.springframework: debug
     org.springframework.security: debug
     io.github.mygoodsupporter: debug
